Converts a natural language query into a 'ggplot2' command.

Converts a natural language query into a 'ggplot2' command string. Queries should be related to the styling of the plot, such as, axis label font size, axis label title, legend, and similar.

gghelp(query = "", print = TRUE)

Arguments

query

Character. A natural language command related to the styling of a ggplot.

print

Boolean. Print out the command or just return it.

Value

Returns a string if there is a matching ggplot command in the database. Otherwise returns NULL.

Details

'gghelp' maintains a database of keywords that match typical queries related to styling 'ggplot2' graphs. Based on the users natural language query, the function tries to find the best match and then returns the ggplot2 command as string.

See also

Examples

gghelp("rotate x-axis labels by 90 degrees")
#> theme(axis.text.x = element_text(angle = 90)) 

gghelp("increase font size on x-axis label")
#> theme(axis.title.x=element_text(size=rel(2))) 

gghelp("set x-axis label to 'Length of Sepal'")
#> xlab('Length of Sepal')